requirements.txt
asgiref==3.3.1
astroid==2.4.1
beautifulsoup4==4.8.2
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.4
django-cogwheels==0.3
django-debug-toolbar==3.2
django-el-pagination==3.3.0
django-filter==2.4.0
django-guardian==2.3.0
django-modelcluster==5.1
django-pygments==0.4
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.12.2
dodgy==0.2.1
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
flake8==3.8.4
flake8-polyfill==1.0.2
html5lib==1.1
httpie==2.4.0
idna==2.10
importlib-metadata==3.3.0
isort==4.3.20
jdcal==1.4.1
l18n==2020.6.1
lazy-object-proxy==1.4.1
Markdown==3.3.3
martor==1.5.8
mccabe==0.6.1
numpy==1.19.4
oauthlib==3.1.0
openpyxl==3.0.5
packaging==20.9
pep8-naming==0.10.0
Pillow==7.2.0
prospector==1.3.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.3
PyJWT==2.0.1
pylint==2.5.3
pylint-celery==0.3
pylint-django==2.1.0
pylint-flask==0.6
pylint-plugin-utils==0.6
pymdown-extensions==8.1
pyparsing==2.4.7
PySocks==1.7.1
python-dotenv==0.15.0
python-http-client==3.3.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.4.1
requests==2.25.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requirements-detector==0.7
rope==0.14.0
selenium==3.141.0
sendgrid==6.5.0
setoptconf==0.2.0
six==1.12.0
snowballstemmer==2.1.0
soupsieve==2.0.1
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
tablib==3.0.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==16.6.0
wagtail==2.10
wagtail-experiments==0.2
wagtail-markdown==0.6
wagtail-metadata==3.4.0
wagtailfontawesome==1.2.1
wagtailmenus==3.0.2
webencodings==0.5.1
Willow==1.4
wrapt==1.11.1
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
zipp==3.4.0
